% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_events.R
\name{search_events}
\alias{search_events}
\title{Requests the Eventful API for event information}
\usage{
search_events(location, date, category, key)
}
\arguments{
\item{location}{A \code{string} of dimension 2 used to denote the integration
region of interest, i.e. [a, b].}

\item{date}{A \code{string} containing the function to be integrated. It
is assumed that \code{x} is used as the variable of interest.}

\item{category}{A \code{string} (integer) used to denote the number of simulations.}

\item{key}{A \code{string} used to control the seed of the random number
generator used by this function.}
}
\value{
A \code{dataframe} containing the following attributes:
\describe{
     \item{I}{Estimated value of the integral}
     \item{var}{Estimated variance of the estimator}
}
}
\description{
This function takes inputs location, date, category, and key.  Its output is a dataframe
containing event information along with information that can be used to create popups with
leaflet.
}
\examples{
mc_int(x_range = c(0,1), fun = "x^2", B = 10^5)
mc_int(x_range = c(0,1), fun = "x^2*sin(x^2/pi)", B = 10^5)
}
\author{
Conner Shoop

Ryan Faris
}
